feat: TipArticle 엔티티 수정
