--- src/main/java/com/sbpb/ddobak/server/domain/tip/entity/TipArticle.java
@@ -1,47 +1,0 @@
-package com.sbpb.ddobak.server.domain.tip.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-/**
- * 계약서 꿀팁 아티클 엔티티
- */
-@Entity
-@Table(name = "tip_articles")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class TipArticle {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "title", nullable = false)
-    private String title;
-
-    @Column(name = "content", nullable = false, columnDefinition = "TEXT")
-    private String content;
-
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Builder
-    public TipArticle(String title, String content, LocalDateTime createdAt) {
-        this.title = title;
-        this.content = content;
-        this.createdAt = createdAt;
-    }
-
-    @PrePersist
-    protected void onCreate() {
-        if (createdAt == null) {
-            createdAt = LocalDateTime.now();
-        }
-    }
-}
