+++ src/main/java/com/sbpb/ddobak/server/domain/tipArticle/entity/TipArticle.java
@@ -1,0 +1,70 @@
+package com.sbpb.ddobak.server.domain.tipArticle.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.type.SqlTypes;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+/**
+ * 꿀팁 아티클 엔티티
+ */
+@Entity
+@Table(name = "tip_articles")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class TipArticle {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "title", nullable = false)
+    private String title;
+
+    @Column(name = "summary", columnDefinition = "TEXT")
+    private String summary;
+
+    @Column(name = "content", columnDefinition = "TEXT")
+    private String content;
+
+    @JdbcTypeCode(SqlTypes.ARRAY)
+    @Column(name = "tags", columnDefinition = "TEXT[]")
+    private List<String> tags;
+
+    @Column(name = "created_at", nullable = false, updatable = false)
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
+    @Builder
+    public TipArticle(String title, String summary, String content, List<String> tags) {
+        this.title = title;
+        this.summary = summary;
+        this.content = content;
+        this.tags = tags;
+    }
+
+    @PrePersist
+    protected void onCreate() {
+        LocalDateTime now = LocalDateTime.now();
+        if (createdAt == null) {
+            createdAt = now;
+        }
+        if (updatedAt == null) {
+            updatedAt = now;
+        }
+    }
+
+    @PreUpdate
+    protected void onUpdate() {
+        updatedAt = LocalDateTime.now();
+    }
+}
